+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/VerticalWell/MapVerticalWell.tsx
@@ -1,5 +1,5 @@
 import { useDroppable } from "@dnd-kit/core";
-import { type ReactNode, forwardRef, useCallback, useMemo } from "react";
+import { useCallback, useMemo } from "react";
 import { t } from "ttag";
 
 import { Ellipsified } from "metabase/core/components/Ellipsified";
@@ -15,6 +15,8 @@ import { removeColumn } from "metabase/visualizer/visualizer.slice";
 import { isDimension, isMetric } from "metabase-lib/v1/types/utils/isa";
 
 import { WellItem } from "../WellItem";
+
+import { WellBox } from "./WellBox";
 
 export function MapVerticalWell() {
   const settings = useSelector(getVisualizerComputedSettings);
@@ -79,7 +81,7 @@ function MapMetricWell() {
   const dispatch = useDispatch();
 
   const { active, isOver, setNodeRef } = useDroppable({
-    id: DROPPABLE_ID.MAP_METRIC,
+    id: DROPPABLE_ID.CANVAS_MAIN,
   });
 
   const isHighlighted = useMemo(() => {
@@ -167,37 +169,3 @@ function MapDimensionWell(props: MapDimensionWellProps) {
     </Box>
   );
 }
-
-interface WellBoxProps {
-  isHighlighted: boolean;
-  isOver: boolean;
-  children: ReactNode;
-}
-
-const WellBox = forwardRef<HTMLDivElement, WellBoxProps>(function WellBox(
-  { children, isHighlighted, isOver },
-  ref,
-) {
-  const borderColor = isHighlighted
-    ? "var(--mb-color-brand)"
-    : "var(--mb-color-border)";
-  return (
-    <Box
-      bg={isHighlighted ? "var(--mb-color-brand-light)" : "bg-light"}
-      p="sm"
-      w="300px"
-      style={{
-        borderRadius: "var(--default-border-radius)",
-        border: `1px solid ${borderColor}`,
-        transform: isHighlighted ? "scale(1.025)" : "scale(1)",
-        transition:
-          "transform 0.2s ease-in-out 0.2s, border-color 0.2s ease-in-out 0.2s, background 0.2s ease-in-out 0.2s",
-        outline:
-          isOver && isHighlighted ? "1px solid var(--mb-color-brand)" : "none",
-      }}
-      ref={ref}
-    >
-      {children}
-    </Box>
-  );
-});
