+++ frontend/src/metabase/visualizer/components/VisualizationCanvas/VerticalWell/PieVerticalWell.tsx
@@ -1,5 +1,5 @@
 import { useDroppable } from "@dnd-kit/core";
-import { type ReactNode, forwardRef, useMemo } from "react";
+import { useMemo } from "react";
 import { t } from "ttag";
 
 import { useDispatch, useSelector } from "metabase/lib/redux";
@@ -16,6 +16,8 @@ import type { DatasetColumn } from "metabase-types/api";
 
 import { WellItem } from "../WellItem";
 
+import { WellBox } from "./WellBox";
+
 export function PieVerticalWell() {
   return (
     <Box>
@@ -31,7 +33,7 @@ function PieMetricWell() {
   const dispatch = useDispatch();
 
   const { active, isOver, setNodeRef } = useDroppable({
-    id: DROPPABLE_ID.PIE_METRIC,
+    id: DROPPABLE_ID.CANVAS_MAIN,
   });
 
   const metric = columns.find(col => col.name === settings["pie.metric"]);
@@ -113,38 +115,3 @@ function PieDimensionWell() {
     </Box>
   );
 }
-
-interface WellBoxProps {
-  isHighlighted: boolean;
-  isOver: boolean;
-  children: ReactNode;
-}
-
-const WellBox = forwardRef<HTMLDivElement, WellBoxProps>(function WellBox(
-  { children, isHighlighted, isOver },
-  ref,
-) {
-  const borderColor = isHighlighted
-    ? "var(--mb-color-brand)"
-    : "var(--mb-color-border)";
-  return (
-    <Box
-      bg={isHighlighted ? "var(--mb-color-brand-light)" : "bg-light"}
-      p="sm"
-      mih="120px"
-      w="300px"
-      style={{
-        borderRadius: "var(--default-border-radius)",
-        border: `1px solid ${borderColor}`,
-        transform: isHighlighted ? "scale(1.025)" : "scale(1)",
-        transition:
-          "transform 0.2s ease-in-out 0.2s, border-color 0.2s ease-in-out 0.2s, background 0.2s ease-in-out 0.2s",
-        outline:
-          isOver && isHighlighted ? "1px solid var(--mb-color-brand)" : "none",
-      }}
-      ref={ref}
-    >
-      {children}
-    </Box>
-  );
-});
