+++ frontend/src/metabase/visualizer/constants.ts
@@ -9,10 +9,8 @@ export const DROPPABLE_ID = {
   X_AXIS_WELL: "X_AXIS_WELL",
   Y_AXIS_WELL: "Y_AXIS_WELL",
 
-  PIE_METRIC: "PIE_METRIC_WELL",
   PIE_DIMENSION: "PIE_DIMENSION_WELL",
 
-  MAP_METRIC: "MAP_METRIC_WELL",
   MAP_DIMENSION: "MAP_DIMENSION_WELL",
   MAP_LATITUDE: "MAP_LATITUDE_WELL",
   MAP_LONGITUDE: "MAP_LONGITUDE_WELL",
